FEAT: Add achievements/badges entities and EF updates
Introduce achievements and badges domain: add Achievement, Badge, BadgeRequirement, AccountAchievement and AccountBadge entities (moved from merged file into separate entity files) and wire them into EF model. Update Reservation includes to eagerly load visible account achievements and taken-out badges. Inject ReservationCacheService into AccountAchievements controller and invalidate reservations cache after toggling achievement visibility or badge taken-out state. Adjust ReservationCacheService cache handling for authenticated reservations. Minor client tweaks: import getTakenOutBadges in useAccountAchievements and simplify ProfileHoverCard empty-badges rendering.
